rev 4; updated 'Unsuccessful' Statistical Analysis
For the prior rev Statistical Analysis sheet, I assumed the definition of "Unsuccessful" to include the outcome for both "failed" and "canceled" campaigns (leading to incorrect values of mean, median, variance, and standard deviation for "Unsuccessful" column J).  This rev, I changed the definition of "Unsuccessful" to only analyze the outcome of "failed" campaigns and updated column J of the sheet accordingly.